Split omnidoc.spec.ts into smaller tests
https://github.com/vitest-dev/vitest/discussions/6511 suggests that vitest 3.x will timeout if a test file takes longer than 60 seconds to complete.
