+++ omnidoc/componentsWithInconsistentCommentsInApiDoc.ts
@@ -1,0 +1,47 @@
+/**
+ * This list is temporary - these components have known inconsistencies,
+ * and we are fixing them. This list will shrink over time as we fix the comments one component at a time.
+ */
+export const componentsWithInconsistentCommentsInApiDoc = [
+  'CartesianAxis',
+  'Cell',
+  'ComposedChart',
+  'Cross',
+  'Curve',
+  'Customized',
+  'Dot',
+  'ErrorBar',
+  'Funnel',
+  'FunnelChart',
+  'Label',
+  'LabelList',
+  'Legend',
+  'Line',
+  'LineChart',
+  'Pie',
+  'PieChart',
+  'PolarAngleAxis',
+  'PolarGrid',
+  'PolarRadiusAxis',
+  'Polygon',
+  'Radar',
+  'RadarChart',
+  'RadialBar',
+  'RadialBarChart',
+  'Rectangle',
+  'ReferenceArea',
+  'ReferenceDot',
+  'ReferenceLine',
+  'ResponsiveContainer',
+  'Sankey',
+  'Scatter',
+  'ScatterChart',
+  'Sector',
+  'Text',
+  'Tooltip',
+  'Trapezoid',
+  'Treemap',
+  'XAxis',
+  'YAxis',
+  'ZAxis',
+];
